Address review comments and fix up .localserver messages
