--- plugins/com.google.cloud.tools.eclipse.appengine.localserver/src/com/google/cloud/tools/eclipse/appengine/localserver/launching/messages.properties
@@ -1,1 +1,0 @@
-UnableToLaunch=Unable to launch App Engine server
