+++ plugins/com.google.cloud.tools.eclipse.appengine.localserver/src/com/google/cloud/tools/eclipse/appengine/localserver/messages.properties
@@ -7,9 +7,23 @@ PORT_OUT_OF_RANGE=Port must be between 0 and 65535.
 
 CREATE_APP_ENGINE_RUNTIME_WIZARD_DESCRIPTION=The App Engine Standard runtime requires the Google Cloud SDK
 CREATE_APP_ENGINE_RUNTIME_WIZARD_TITLE=App Engine Standard Runtime
+SERVICES_HAVE_SAME_ID="{0}" and "{1}" have same App Engine Service ID: {2}
 OPEN_CLOUD_SDK_PREFERENCE_BUTTON=Open the Cloud SDK Location preference page when the wizard closes
 RUNTIME_WIZARD_CLOUD_SDK_NOT_FOUND=Cannot find the Google Cloud SDK.
 
 SERVER_STARTING_TEMPLATE=<starting...> {0}
 SERVER_STOPPING_TEMPLATE=<stopping...> {0}
 SERVER_STOPPED_TEMPLATE=<stopped> {0}
+
+CANNOT_DETERMINE_EXECUTION_CONTEXT=Cannot determine server execution context
+NO_MODULE_FOUND_FOR=no module found for {0}
+RUNNING_SERVER_SHOULD_HAVE_A_LAUNCH=Running server should have a launch
+SERVER_ALREADY_RUNNING=Server is already running
+SERVER_ALREADY_RUNNING_IN_MODE=Server is already running in "{0}" mode
+UNABLE_TO_LAUNCH=Unable to launch App Engine server
+
+cloudsdk.not.configured=Could not run project because the Cloud SDK was not found. \
+ Check Preferences > Google Cloud Tools > SDK location and try again.
+cloudsdk.out.of.date=Could not run project because the installed Cloud SDK is too old. \
+ Run `gcloud components update` and try again.
+
