+++ plugins/com.google.cloud.tools.eclipse.appengine.localserver/src/com/google/cloud/tools/eclipse/appengine/localserver/server/LocalAppEngineServerBehaviour.java
@@ -33,7 +33,6 @@ import com.google.common.annotations.VisibleForTesting;
 import java.io.File;
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -220,8 +219,8 @@ public class LocalAppEngineServerBehaviour extends ServerBehaviourDelegate
         logger.log(Level.INFO, attribute + ": port " + port + " in use. Picking an unused port.");
         port = 0;
       } else {
-        throw new CoreException(newErrorStatus(
-            MessageFormat.format(Messages.getString("PORT_IN_USE"), String.valueOf(port))));
+        throw new CoreException(
+            newErrorStatus(Messages.getString("PORT_IN_USE", String.valueOf(port))));
       }
     }
     return port;
