--- plugins/com.google.cloud.tools.eclipse.appengine.localserver/src/com/google/cloud/tools/eclipse/appengine/localserver/server/messages.properties
@@ -1,5 +1,0 @@
-cloudsdk.not.configured=Could not run project because the Cloud SDK was not found. \
- Check Preferences > Google Cloud Tools > SDK location and try again.
-cloudsdk.out.of.date=Could not run project because the installed Cloud SDK is too old. \
- Run `gcloud components update` and try again.
-
